feat(google/map): remove callbacks when calling fitbounds
